+++ src/vatensor/linalg.cpp
@@ -6,66 +6,36 @@
 #include <vector>                 // for vector
 #include "reduce.hpp"               // for sum
 #include "varray.hpp"      // for VArray, VArrayTarget, VScalar, axes...
-#include "vassign.hpp"     // for assign
 #include "vmath.hpp"                // for multiply
 #include "xtensor/xslice.hpp"     // for all, ellipsis, newaxis, xall_tag
 
-// struct Dot {
-// 	template <typename GivenAxes, typename A, typename B>
-// 	auto operator()(GivenAxes&& axes, A&& a, B&& b) const {
-// 		auto prod = std::forward<A>(a) * std::forward<B>(b);
-// 		return xt::sum(prod, std::forward<GivenAxes>(axes), std::tuple<xt::evaluation_strategy::lazy_type>());
-// 	}
-//
-// 	template <typename A, typename B>
-// 	inline auto operator()(A&& a, B&& b) const {
-// 		auto prod = std::forward<A>(a) * std::forward<B>(b);
-// 		return xt::sum(prod, std::tuple<xt::evaluation_strategy::lazy_type>());
-// 	}
-// };
-
-va::VScalar va::reduce_dot(const VArray& a, const VArray& b) {
-	std::shared_ptr<va::VArray> prod_cache;
-	va::multiply(&prod_cache, a, b);
-	return sum(*prod_cache);
-}
-
-void va::reduce_dot(VArrayTarget target, const VArray& a, const VArray& b, const axes_type& axes) {
-	std::shared_ptr<va::VArray> prod_cache;
-	va::multiply(&prod_cache, a, b);
-	va::sum(target, *prod_cache, axes);
-
-	// TODO This doesn't work because prod or a and b are lost, either way it crashes.
-	// The upside to the above implementation is that no additional code is generated.
-	// But it's also a bit slower than if it was fully lazy and accelerated, probably.
-	// va::xreduction_inplace<promote::num_matching_float_or_default<double_t>>(
-	// 	NormL0{}, axes, target, array.read
-	// );
-}
-
 void va::dot(VArrayTarget target, const VArray& a, const VArray& b) {
-	if (a.dimension() == 1 && b.dimension() == 1) {
-		std::shared_ptr<va::VArray> prod_cache;
-		va::multiply(&prod_cache, a, b);
-		va::assign(target, va::sum(*prod_cache));
+	if (a.dimension() == 0 || b.dimension() == 0) {
+		va::multiply(target, a, b);
+		return;
 	}
-	else if (a.dimension() == 2 && b.dimension() == 2) {
-		return va::matmul(target, a, b);
+	if (a.dimension() <= 2 && b.dimension() <= 2) {
+		va::matmul(target, a, b);
+		return;
 	}
-	else if (a.dimension() == 0 || b.dimension() == 0) {
-		return va::multiply(target, a, b);
-	}
-	else if (b.dimension() == 1) {
-		std::shared_ptr<va::VArray> prod_cache;
-		va::multiply(&prod_cache, a, b);
-		va::sum(target, *prod_cache, std::vector { static_cast<std::ptrdiff_t>(-1) });
-	}
-	else {
-		throw std::runtime_error("tensordot is not yet implemented");
-	}
+
+	throw std::runtime_error("tensordot is not yet implemented");
 }
 
 void va::matmul(VArrayTarget target, const VArray& a, const VArray& b) {
+	if (a.dimension() == 0 || b.dimension() == 0) {
+		throw std::runtime_error("matmul does not accept scalars");
+	}
+	if (b.dimension() == 1) {
+		va::reduce_dot(target, a, b, {-1});
+		return;
+	}
+	if (a.dimension() == 1) {
+		const auto promoted_a = a.sliced({xt::all(), xt::newaxis()});
+		va::reduce_dot(target, *promoted_a, b, {-2});
+		return;
+	}
+
 	const std::shared_ptr<VArray> a_broadcast = a.sliced({ xt::ellipsis(), xt::newaxis() });
 	const std::shared_ptr<VArray> b_broadcast = b.sliced({ xt::ellipsis(), xt::newaxis(), xt::all(), xt::all() });