+++ src/vatensor/linalg.hpp
@@ -5,9 +5,6 @@
 #include "varray.hpp"
 
 namespace va {
-	VScalar reduce_dot(const VArray& a, const VArray& b);
-	void reduce_dot(VArrayTarget target, const VArray& a, const VArray& b, const axes_type& axes);
-
 	void dot(VArrayTarget target, const VArray& a, const VArray& b);
 	void matmul(VArrayTarget target, const VArray& a, const VArray& b);
 }